Create READ.ME
This project is a DEMO of some basic SWIFT and Xcode functionality.   It assumes you have some familiarity with Xcode, and know how to set up a simple UI which consists of an image view and five textFields.   Project Mongoose covers topics you should know before starting this Demo.

A PDF of a Keynote is attached and can be used as a guide in going through the Demo.   Extensive documentation and references to specific slides in the Keynote are in the code.

This demo can be freely copied by persons learning Swift.   It may not be copied or used by commercial organizations without prior written arrangements with Quantum Films Software.